changed storage system for tiddlers in browser
The tiddlers and bags are stored in a prived "store" object. they can be
accessed with getTiddler, getBag, and each.

The new store object allows storage of more than one tiddler with the
same title by splitting tiddlers up by bag. Basically as follows:

	store = {
		bagName1: {
			thing: tiddlyweb.Bag(),
			tiddlers: {
				tiddlerName1: tiddlyweb.Tiddler(),
				tiddlerName2: tiddlyweb.Tiddler(),
			}
		},
		bagName2: {
			thing: tiddlyweb.Bag(),
			tiddlers: {
				tiddlerName1: tiddlyweb.Tiddler(),
				tiddlerName2: tiddlyweb.Tiddler(),
			}
		},
	}

The pending list and recipe object currently still exist, and are still
public. This is likely to change in the future.