Added jQuery-like filtering of tiddlers
The Store object is now a function that when called (with optional
filters) returns a list of tiddlers that match the optional filter (or
all tiddlers). This list can then be acted upon with further
filters/functions (influenced by TidQL (http://tidql.tiddlyspace.com)).

e.g.:

	var store = tiddlyweb.Store().refreshTiddlers();
	var coreNotMissing = store('title',
	'TiddlySpace').tag('excludeMissing').reduce('', function(tid, str) {
		return str + tid.title + ' '; }).replace(/ $/, '');

	console.log(coreNotMissing);

prints a space separated list of tiddler titles that include
"TiddlySpace" and are tagged excludeMissing.