Added some tests and fixtures. Also added a default bag
When using addTiddler, chrjs-store attempts to add a default bag (based
on the current space) if one is not present in the tiddler already. This
fixes a bug where that default bag hasn't been used before, so is not
present in the store. In this case, it now creates a new bag.