Modified the Tiddlers object to remove prototypal inheritance
Previously, I was using prototypal inheritance to inherit an Array,
which I could then happily extend to create a new Array-like object
without modifying Array.prototype itself.

It turns out though that several functions of Arrays don't work when you
do this:

* The array.length property correctly keeps track when using push, etc
* The automatic array extending (e.g. a = [1,2,3]; a[400] = 4; console.log(a.length)) doesn't work.
* The array.concat function doesn't work
* probably some more that I didn't test don't work either.

It runs out that most stuff works fine, hence why I didn't come across
this issue before. The code that highlighted this issue in Mustache.js,
which does:

	Object.prototype.toString.call(myArray) === '[object Array]'

to check whether the object is an array (jQuery also does this).

The Tiddlers object is now a function, that creates a new, blank, array
and extends it directly with the helper functions, which seems to work,
but doesn't seem quite as clean.