added a lightweight RequireJS object
RequireJS defers oading any scripts until after all scripts in the page
have loaded by putting everything in a setTimeout. As Chrjs-Store is a
library, it is likely to be required by other scripts in the page. In
order for this to happen, one of two things must happen:
	1. The scripts requiring it must also use RequireJS
	2. RequireJS must evaluate modules and requires as it finds them,
	instead of after everything else.

As a library, I don't want to force people to use RequireJS if they
don't want to, they should just be able to use it as with any other
library. THis means I've gone with option 2.

To satisfy this, I've modified a similar lightweight object from the ACE
project
(https://github.com/mozilla/ace/blob/master/build_support/mini_require.js).