Changed filter.fn.bind to bind to filters instead of tiddlers
Using the new parser code in filter-syntax, the bind function now binds
a function created from the AST of the current set of filters
(createTester) to the callback function supplied.

The upshot of this is that calling, for example:

	ts('tag', 'foo').space('bar).bind(function(tiddler) {
		// do stuff here
	});

Will cause the bound function to be called every time a tiddler tagged
"foo" in space "bar" is created/modified/deleted.

Previously, the bind function only bound to tiddlers currently in the
filtered list, so new tiddlers matching would not be picked up.