Fixed a few bugs:
The mini-require in chrjs-store was interfering with the mini-require in
ace. This chrjs-store specific version is now contained to
chrjs-store.js only so doesn't mess anything up

There was a bug in space causing it to fail when the recipe wasn't
present. This now doesn't happen, and space continues to work inside the
bind function if the recipe loads up between the two happening.

There was a minor bug in the event  trigger function.