Released version 0.3.0
Additions include a new filter syntax:

	ts()find('#tag +modifier @space [field=value], [[OrTitle]] "text
	match"')

Syntax is loosely based on CSS selectors (space means AND, comma means
OR), and is as follows:
* [[Title]]
* #tag
* @space
* +modifier
* [field=value]
* "text"
* !match
Not also works with the other options, e.g. !#tag, [field!=value], etc.

The other addition is the ability to bind to filters, rather than
individual tiddlers, meaning any tiddler matching the given filter will
be triggered.

For example:

	ts('title', 'foo').find('#tag +modifier,
	[field=value]').space().bind(function(tiddler) {
		// bind to the filters given up until this point
	});