Fixed a bug in firefox with localStorage handling
The localStorage object doesn't work quite like arrays or (normal)
objects.

In firefox at least, you can't iterate over it. You need to get the
length, and iterate up to that, calling the key() method for each number
in ordder to get the actual key.

This issue doesn't show up in Chrome, which allows you to iterate over
localStorage as if it were an array.