Modified add so that space doesn't need to be present if no bag is
The add function doesn't require the tiddler to have a bag. It gets
around this by setting the bag by default to the public bag of the space
using getSpace. However, this means that the tiddler doesn't get added
until the space is figured out, which is potentially a problem if the
user requires the tiddler immediately.

This change adds the tiddler first, without a bag, and then tries to
figure out the bag afterwards, so that the tiddler always exists.

This now means that tiddlers in the store do not necessarily always have
a bag. This is the reason for the changes in other areas.