Made lastSync optional and added makeCopy function
This means that localStore can now be used for both the synced store and
the pending store (which I've renamed modified and is no longer public).
The deep copy of tiddlers also now happens inside localStore. This all
means that a lot of cruft can be removed from store.js, which is a good
thing.