renamed getSpace to getDefaults and separated out defaultContainer logic
You can now pass in a (set of) default container(s) to the store, and
have the store use it/them as defaults instead of assuming a TiddlySpace
style public recipe and public bag.

If you don't pass this in, then it will try to determine a sensible
default, based upon what it finds at "/". If it recognises TiddlySpace
style public/private recipes, etc, then it will default to them.
Otherwise it will use whatever recipe/bag (in that order) it finds. If
there are no tiddler(s) present at "/", then you will need to supply
your own default container.

defaultContainer accepts the following:

	defaultContainer = new tiddlyweb.Recipe(name, host);
	defaultContainer = new tiddlyweb.Bag(name, host);

or, an object like the following:

	defaultContainer = {
		pullFrom: <chrjs-container>,
		pushTo: <chrjs-container>
	};

where pushTo is the default container that you want to save tiddlers to
and pullFrom is the default container that you want to refresh the store
from.

The new getDefaults function (renamed from getSpace (which is no longer
available)), passes an object with these properties into the callback.