Released version 0.4.0
Changes include an option to specify custom default containers, and
somechanges under the hood that should mean you can add more than one
locally modified tiddler with the same title to the store, as long as
they have separate bags. Also removed the pending object from the public
interface.

Previously, it existed only because detecting the difference was overly
complicated, however, with the .dirty() filter, that is no longer the
case.

This version should also now be a lot more suitable for non-TiddlySpace
use cases (it still requires TiddlyWeb though)