Fixed some regressions around binding
There was a bug in the triggering code on save, meaning that the check
on whether to trigger an event upon a modified tiddler being added was
incorrect.

With synced tiddlers, this happens by comparing revision numbers, which
we assume to be relatively accurate (coming from the server). But with
modified tiddlers, we cannot rely on this, so I've written a compare
function that compares tiddler equality directly. THis should make the
checking code (and thus triggering) more correct.

There was also a bug in the filtering code caused by not setting the
store.recipe attribute upon getDefaults (getSpace previously did this).
Ideally this should not be needed, but some filters require synchronous
access to such things, which getDefaults cannot provide sufficiently.