Changed save callback to compare with preSave version
There was a bug in the compare section of the saveTiddler callback. When
the tiddler has been saved, the server can add certain attributes onto
the tiddler (created, _hash, etc), meaning that the comparison with the
current tiddler in modified can fail. By changing this to compare with
preSave instead, we ensure that we're only checking whether the tiddler
has been updated whilst saving, and not that they're the same thing
(which is more correct).

Also added an extra test to the Save tiddler test that passes a tiddler
into the save function directly.