Extracted parent/child stuff into a separate module
After a discussion with @jermolene, we decided that the parent/child
stuff, while important, isn't appropriate for core chrjs-store, and that
it should be a plugin instead.

It now lives in the extensions folder, complete with all the tests
associated with it.

Also, the Constructor for the filtering stuff can now be accessed
directly from the store, e.g.:

	var store = tiddlyweb.Store();
	var emptyCollection = store.Collection(store, []);

You can use the new tree-filters module by doing:

	var store = tiddlyweb.Store();
	$.extend(store.fn, require('tree-filters'));

which will extend the default set of filters to include the tree
specific ones