Released version 0.5.0
Changes include new sort and limit filters and the exposure of the
filter collection constructor function.

Use the sort function either like the inbuilt Array sort, or pass in a
string consisting of a comma separated list of field names to sort by
(add a minus sign to sort in desc order).

The Collection object for filtering can now be created directly (useful
if you want to manually put tiddlers into it and want to bypass the
store). Call it with:

	var store = tiddlyweb.Store();
	var list = store.Collection(tiddlers);

Where tiddlers is optional and a list of tiddlers that you want the
collection to contain.