Refactored complicated argument handling
Added error functions and a function that takes in an object or string
and returns an object that contains the tiddler it references, along
with the modified version, it's title, the raw tiddler that was passed
in. This is because there was previously a lot of cruft at the top of
each function that attempted to deal with the dynamic parameter list.
This removes that cruft and should make the resultant code a lot easier
to read.

The store find function is now one line and simply defers to another
function. This is primarily so that it doesn't take up too much space in
the console when being logged in browsers.