Breaking Changes: Made map and reduce more ES5-like
Previously, .map essentially did a map and filter operation in one. The
new behaviour is closer to ES5 syntax, and uses the same API. The
difference being that a proper tiddler collection is returned. The
reduce function also now copies the ES5 API.

This has has been done to be more compatible with people's expectations.

I've also added a filter function, which takes the place of the old map
functionality that could be used in that way.

With the new filter function in place it also makes sense to borrow the
tester functions from the filter module where available so that they are
not duplicated. This also does that.

As a result some old behaviour is now broken. Check the test_filter diff
in this commit to find out what's changed.