Breaking Change: split remove into remove and destroy
remove is now simply the inverse of add: it removes locally added
tiddlers.

destroy is the inverse of save: it deletes the tiddler from both the
server, and locally.

If all you want is to permanently get rid of a tiddler, then use
destroy. If you only want to revert to the server version, use remove.