Released version 0.6.0
This is quite a large release as it includes serveral breaking changes:

* Two of the modules have been renamed:
	filter -> collection
	filter-syntax -> filter
* the map and reduce functions in the collection module (nee filter),
have been changed to match the ES5 spec more closely. This mean that
there is now an extra function: filter, which performs the filter part
of the old map function
* The refresh code has been split out into a separate module. This means
that searches will also be refreshed (in addition to the default recipe)
when calling refresh. This should make it easier to add extra container
types and localStores in the future (a shadow like store is planned for
adding pseudo-tiddlers from, for example, the DOM).
* The remove function has been split out into two separate functions:
	remove: function(tiddler) { ... }
		removes from local/modified only
	destroy: function(tiddler, callback) { ... }
		removes all references both locally and on the server

cc: @jdlrobson