Fixed a coumple of regressions
There was an issue in the save function where the tiddler in modified
was taking precedence over the tiddler acutally passed into the
function.

There was also an issue with bootstrapping the library, with the wrong
arguments being passed into almond.js.

There was also an issue with the refresh code being called before
getDefaults returned, meaning it was refreshing before it had any
containers to refresh.