Fixed a bug in multiple store implementation
There was a problem with the queue implementation in host.js where the
callback currently being fired was not removed until after it had been
completed. This meant that if getDefaults was called inside that
callback, the same callback would be run again, resulting in an infinite
loop. This code now removes the callback before running it.